New address now needs a secret
The new address endpoint is now a POST and needs a "secret" in the body.
The secret is used to create a hash that will be store in the DB, this can be used when authenticating users.

Bcrypt files are currently committed. Will set a up a pipeline in the future